Optimise the save format for quicker compression and shorter writes
THIS COMMIT BREAKS COMPATIBILITY WITH SAVES FROM OTHER VERSIONS OF
MERITOUS.

For compressibility, the XOR encoding is dropped. The format of 32-bit
integers written and read by functions FWInt and FRInt is also changed
to be a variable-length format that takes up 1 or 2 bytes in the most
common cases, and up to 5 in uncommon cases.